Refactor module file names and directories
