+++ src/basestation/CMakeLists.txt
@@ -1,14 +1,14 @@
 set(MODULE_HDR
 	modules/console.hpp
-	modules/controller_config.hpp
-	modules/controller_calibration_popup.hpp
-	modules/controller_binding_popup.hpp
+	modules/input_config/controller_config.hpp
+	modules/input_config/controller_calibration_popup.hpp
+	modules/input_config/controller_binding_popup.hpp
 )
 set(MODULE_SRC
 	modules/console.cpp
-	modules/controller_config.cpp
-	modules/controller_calibration_popup.cpp
-	modules/controller_binding_popup.cpp
+	modules/input_config/controller_config.cpp
+	modules/input_config/controller_calibration_popup.cpp
+	modules/input_config/controller_binding_popup.cpp
 )
 
 set(WIDGET_HDR
